PP-3518 fix: after review
