+++ tests/test_signature_handler.py
@@ -1,5 +1,6 @@
 
 from payment_page_sdk.signature_handler import SignatureHandler
+from payment_page_sdk.callback import Callback
 import copy
 import unittest
 
@@ -36,3 +37,29 @@ class SignatureHandlerTest(unittest.TestCase):
             self.payment_data['body']['signature'],
             self.signature_handler.sign(compare_data['body'])
         )
+
+    def test_multilevel_callback(self):
+        signature_handler = SignatureHandler(self.secret)
+        payment_data_raw = '{' +\
+            '"payment": {' +\
+                '"id": "test-payment"' +\
+            '},' +\
+            '"errors": [' +\
+                '{' +\
+                    '"code": "123",' +\
+                    '"message": "grand crash",' +\
+                    '"description": [' +\
+                        '"description-str",' +\
+                        '{' +\
+                            '"description1": 1' +\
+                        '}' +\
+                    ']' +\
+                '},' +\
+                '{' +\
+                    '"code": "456",' +\
+                    '"message": "minor crash"' +\
+                '}' +\
+            '],' +\
+            '"signature": "i6F9q/cfj8T+r/9z1U1f+WgGe0Y/2L2Ml7pp6x6GKLmS0SwbEzcD1/DJ2Hx4PwSgyD696SU19MjtHBF0gv+LKA=="' +\
+        '}'
+        Callback(payment_data_raw, signature_handler)
